src/test/java/com/restaurant/Repository/TablesRepositoryTest.java
package com.restaurant.Repository;

import com.restaurant.Entity.Tables;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;


@RunWith(SpringRunner.class)
@DataJpaTest
@TestPropertySource(properties = {
        "spring.jpa.hibernate.ddl-auto=validate"
})
class TablesRepositoryTest {

    @Autowired
    private TablesRepository repo;

    @Test
    void canFindAll() {


    }

    @Test
    void findMaxId() {
    }

    @Test
    void findTableById() {
    }

    @Test
    void findAvailableTables() {
    }
}